HW_2_CashMachine_v2. Поменял константы с int на double

+++ src/main/java/org/example/hw_2_cashMachine/model/constants/Constants.java
@@ -1,16 +1,16 @@
 package org.example.hw_2_cashMachine.model.constants;
 
 public class Constants {
-    public final int WITHDRAWAL_LIMIT = 30000;
-    public final int AMOUNT_REASON_FOR_CASHBACK = 1000000;
-    public final double CASHBACK_AMOUNT = 2000;
-    public final int DEBET_SCORE_OPERATIONS_BLOCK_LIMIT = -20000;
+    public final double WITHDRAWAL_LIMIT = 30000.00;
+    public final double AMOUNT_REASON_FOR_CASHBACK = 1000000.00;
+    public final double CASHBACK_AMOUNT = 2000.00;
+    public final double DEBET_SCORE_OPERATIONS_BLOCK_LIMIT = -20000.00;
 
-    public int getWITHDRAWAL_LIMIT() {
+    public double getWITHDRAWAL_LIMIT() {
         return WITHDRAWAL_LIMIT;
     }
 
-    public int getAMOUNT_REASON_FOR_CASHBACK() {
+    public double getAMOUNT_REASON_FOR_CASHBACK() {
         return AMOUNT_REASON_FOR_CASHBACK;
     }
 
@@ -18,7 +18,7 @@ public class Constants {
         return CASHBACK_AMOUNT;
     }
 
-    public int getDEBET_SCORE_OPERATIONS_BLOCK_LIMIT() {
+    public double getDEBET_SCORE_OPERATIONS_BLOCK_LIMIT() {
         return DEBET_SCORE_OPERATIONS_BLOCK_LIMIT;
     }
 }
